extract getPrice into own file
