+++ src/components/ProductList/index.js
@@ -1,12 +1,6 @@
 import React from 'react';
 import './ProductList.css';
-
-function getPrice(priceFloat) {
-  const overOne = parseInt(priceFloat, 10);
-  const belowOne = Math.round((priceFloat - overOne) * 100);
-  const belowText = belowOne < 10 ? `0${belowOne}` : belowOne;
-  return `kr ${overOne},${belowText}`;
-}
+import getPrice from '../../utils/getPrice';
 
 function ProductListItem(props) {
   const { product } = props;
